refactor(agents): KG-812. Move lock and exception utilities to agents-utils module
