+++ a2a/a2a-core/build.gradle.kts
@@ -20,6 +20,7 @@ kotlin {
             dependencies {
                 api(libs.kotlinx.serialization.json)
                 api(libs.kotlinx.coroutines.core)
+                implementation(project(":agents:agents-utils"))
             }
         }
