+++ a2a/a2a-server/build.gradle.kts
@@ -18,6 +18,7 @@ kotlin {
                 api(libs.ktor.client.core)
                 api(libs.ktor.client.content.negotiation)
                 api(libs.ktor.serialization.kotlinx.json)
+                implementation(project(":agents:agents-utils"))
                 implementation(libs.oshai.kotlin.logging)
             }
         }
