+++ agents/agents-core/src/commonMain/kotlin/ai/koog/agents/core/feature/writer/FeatureMessageFileWriter.kt
@@ -2,7 +2,7 @@ package ai.koog.agents.core.feature.writer
 
 import ai.koog.agents.core.feature.message.FeatureMessage
 import ai.koog.agents.core.feature.message.FeatureMessageProcessor
-import ai.koog.agents.core.utils.MutexCheck.withLockCheck
+import ai.koog.agents.lock.withLockCheck
 import io.github.oshai.kotlinlogging.KotlinLogging
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
