+++ agents/agents-utils/build.gradle.kts
@@ -12,7 +12,9 @@ kotlin {
     sourceSets {
         commonMain {
             dependencies {
+                implementation(libs.kotlinx.coroutines.core)
                 implementation(libs.kotlinx.serialization.json)
+                implementation(libs.oshai.kotlin.logging)
             }
         }
