+++ agents/agents-utils/src/commonMain/kotlin/ai/koog/agents/exception/CancellationExceptionExt.kt
@@ -1,0 +1,18 @@
+package ai.koog.agents.exception
+
+import kotlinx.coroutines.CancellationException
+
+/**
+ * Extension property that retrieves the root cause of a [CancellationException].
+ *
+ * This is used in cases when non-suspended code launches a coroutine that raises exceptions.
+ * The method helps to get the root exception causing the outer coroutine cancellation.
+ */
+public val CancellationException.rootCause: Throwable?
+    get() {
+        var rootCause: Throwable? = this
+        while (rootCause != null && rootCause is CancellationException) {
+            rootCause = rootCause.cause
+        }
+        return rootCause
+    }
