+++ agents/agents-utils/src/commonMain/kotlin/ai/koog/agents/lock/RWLock.kt
@@ -1,21 +1,18 @@
-package ai.koog.a2a.utils
+package ai.koog.agents.lock
 
-import ai.koog.a2a.annotations.InternalA2AApi
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
 
-// FIXME copied from agents-core module, because a2a does not depend on other Koog modules.
-//  Do we want to make a global utils module for cases like this?
 /**
  * A Kotlin Multiplatform read-write lock that allows concurrent read access while guaranteeing
  * exclusive write access.
  *
  * The implementation uses two [Mutex] instances:
  * - [writeMutex] is held while a writer runs, or while there is at least one active reader.
- *   The first reader acquires it and the last reader releases it.
+ *   The first reader acquires it, and the last reader releases it.
  * - [readersCountMutex] protects [readersCount] so that reader acquisition/release is atomic.
  *
- * Concurrency caveats:
+ * Concurrency warnings:
  *
  * 1. **Not reentrant.** The underlying [Mutex] has no notion of an owning coroutine, and this
  *    lock does not track owners either. Calling [withWriteLock] from inside another
@@ -34,7 +31,6 @@ import kotlinx.coroutines.sync.withLock
  * 5. **Suspend-only.** There is no blocking or `tryLock` API; this lock is usable only from
  *    coroutine code.
  */
-@InternalA2AApi
 public class RWLock {
     private val writeMutex = Mutex()
     private var readersCount = 0
@@ -43,7 +39,7 @@ public class RWLock {
     /**
      * Run the given [block] of code while holding the read lock. Multiple coroutines may run
      * their read blocks concurrently. See the class-level KDoc for reentrancy and cancellation
-     * caveats.
+     * warnings.
      */
     public suspend fun <T> withReadLock(block: suspend () -> T): T {
         readersCountMutex.withLock {
@@ -67,7 +63,7 @@ public class RWLock {
      * Run the given [block] of code while holding the write lock. Only one coroutine at a time
      * may run under the write lock, and no readers are active while it runs. Must not be called
      * from a coroutine that already holds the read or write lock on this instance — see the
-     * class-level KDoc for reentrancy caveats.
+     * class-level KDoc for reentrancy warnings.
      */
     public suspend fun <T> withWriteLock(block: suspend () -> T): T {
         writeMutex.withLock {
