+++ prompt/prompt-llm/build.gradle.kts
@@ -12,9 +12,9 @@ kotlin {
     sourceSets {
         commonMain {
             dependencies {
+                api(project(":agents:agents-utils"))
                 api(libs.kotlinx.serialization.core)
                 api(libs.jetbrains.annotations)
-                api(project(":agents:agents-utils"))
             }
         }
     }
